Schema Refinement for `Card` and `Mechanic`
- Refine the schemas of `Card` and `Mechanic` for crawler development
- Add more unit tests for `Card` and `Mechanic`

+++ core/src/main/java/com/hearthintellect/dao/MechanicRepository.java
@@ -4,7 +4,7 @@ import com.hearthintellect.model.Mechanic;
 
 public interface MechanicRepository {
 
-    Mechanic findById(long mechanicId);
+    Mechanic findById(int mechanicId);
 
     Iterable<Mechanic> findAllByName(String name);
