+++ core/src/main/java/com/hearthintellect/dao/mongo/MechnicRepositoryImpl.java
@@ -21,7 +21,7 @@ public class MechnicRepositoryImpl implements MechanicRepository {
     }
 
     @Override
-    public Mechanic findById(long mechanicId) {
+    public Mechanic findById(int mechanicId) {
         return createQuery().field("_id").equal(mechanicId).get();
     }
